# demo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "email": "yinjie_luo@intsig.net"
        }
    },
    "host": "http://192.168.6.50:10006",
    "basePath": "/edapi/bankbills/",
    "paths": {
...
        "/edapi/bankbills/analyse/lsyz/completeness/account": {
            "get": {
                "description": "流水验证-数据完整性-账号信息",
                "tags": [
                    "流水验证"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.LsyzCAccountResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/lsyz/completeness/balance": {
            "get": {
                "description": "流水验证-数据完整性-余额完整",
                "tags": [
                    "流水验证"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "minimum": 5,
                        "type": "integer",
                        "description": "默认5条/页",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "第一页开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.LsyzBalanceResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/lsyz/completeness/summary": {
            "get": {
                "description": "流水验证-数据完整性-摘要完整",
                "tags": [
                    "流水验证"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "minimum": 5,
                        "type": "integer",
                        "description": "默认5条/页",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "第一页开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.LsyzSummaryResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/edapi/bankbills/project/add": {
            "get": {
                "description": "项目管理-新建项目",
                "tags": [
                    "项目管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目备注",
                        "name": "note",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "项目类型 1:企业 2:个人 3:其他",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.AddProjectResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/project/delete": {
            "get": {
                "description": "项目管理-删除指定项目",
                "tags": [
                    "项目管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "待删除项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.Result1"
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/project/query": {
            "get": {
                "description": "项目管理-项目列表",
                "tags": [
                    "项目管理"
                ],
                "parameters": [
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "项目创建时间排序搜索项 0:默认倒序 1:正序",
                        "name": "date_desc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间 秒级时间戳",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键字搜索 包含项目名称和备注的模糊匹配",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "minimum": 5,
                        "type": "integer",
                        "description": "默认5条/页",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "第一页开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间 秒级时间戳",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "项目类型搜索项 1:企业 2:个人 3:其他",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.ProjectListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/project/update": {
            "get": {
                "description": "项目管理-更新项目",
                "tags": [
                    "项目管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目备注",
                        "name": "note",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "项目类型 1:企业 2:个人 3:其他",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.Result1"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "rest.AbnormalTradeI": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "交易账号",
                    "type": "string"
                },
                "ending_balance": {
                    "description": "期末余额",
                    "type": "integer"
                },
                "initial_balance": {
                    "description": "初期余额",
                    "type": "integer"
                },
                "name": {
                    "description": "交易户名",
                    "type": "string"
                },
                "summary": {
                    "description": "交易摘要",
                    "type": "string"
                },
                "trade_amount": {
                    "description": "交易金额",
                    "type": "integer"
                },
                "trade_time": {
                    "description": "交易时间",
                    "type": "string"
                }
            }
        },
        "rest.AbnormalTradeI1": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "交易账号",
                    "type": "string"
                },
                "ending_balance": {
                    "description": "期末余额",
                    "type": "integer"
                },
                "name": {
                    "description": "交易户名",
                    "type": "string"
                },
                "summary": {
                    "description": "交易摘要",
                    "type": "string"
                },
                "trade_amount": {
                    "description": "交易金额",
                    "type": "integer"
                },
                "trade_time": {
                    "description": "交易时间",
                    "type": "string"
                }
            }
        },
...
        "rest.AddProjectResp": {
            "type": "object",
            "properties": {
                "project_id": {
                    "description": "新建成功之后的项目id",
                    "type": "integer"
                }
            }
        },
...
        "rest.AssociateInfoI": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "银行账户",
                    "type": "string"
                },
                "name": {
                    "description": "隶属银行",
                    "type": "string"
                }
            }
        },
...
        "rest.BalanceTypeI": {
            "type": "object",
            "properties": {
                "timer": {
                    "description": "2023-01",
                    "type": "string"
                },
                "type": {
                    "description": "1 缺失 2 错误 3 正确",
                    "type": "integer"
                }
            }
        },
...
        "rest.LsyzBalanceResp": {
            "type": "object",
            "properties": {
                "abnormal_trade": {
                    "description": "异常记录",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.AbnormalTradeI"
                    }
                },
                "balance_type": {
                    "description": "月余额类型",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.BalanceTypeI"
                    }
                },
                "error_num": {
                    "description": "月错误总月数",
                    "type": "integer"
                },
                "interval_type": {
                    "description": "区间类型 0月 1周",
                    "type": "integer"
                },
                "total_num": {
                    "description": "余额总月数",
                    "type": "integer"
                }
            }
        },
        "rest.LsyzCAccountResp": {
            "type": "object",
            "properties": {
                "none_same_name_account": {
                    "description": "非同名账户信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.NSMInfo"
                    }
                },
                "theme_account": {
                    "description": "主体账户",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ThemeAccountInfo"
                    }
                }
            }
        },
        "rest.LsyzSummaryResp": {
            "type": "object",
            "properties": {
                "abnormal_trade": {
                    "description": "异常记录",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.AbnormalTradeI1"
                    }
                },
                "error_num": {
                    "description": "摘要缺失月数",
                    "type": "integer"
                },
                "interval_type": {
                    "description": "区间类型 0月 1周",
                    "type": "integer"
                },
                "summary_type": {
                    "description": "月摘要类型",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.SummaryTypeI"
                    }
                },
                "total_num": {
                    "description": "摘要总月数",
                    "type": "integer"
                }
            }
        },
...
        "rest.NSMInfo": {
            "type": "object",
            "properties": {
                "trade_account": {
                    "description": "交易账号",
                    "type": "string"
                },
                "trade_name": {
                    "description": "交易户名",
                    "type": "string"
                },
                "trade_period": {
                    "description": "交易时间段",
                    "type": "string"
                },
                "transfer_in": {
                    "description": "向同名账户转入 单位元",
                    "type": "integer"
                },
                "transfer_in_times": {
                    "description": "向同名账户转入次数",
                    "type": "integer"
                },
                "transfer_out": {
                    "description": "向同名账户转出 单位元",
                    "type": "integer"
                },
                "transfer_out_times": {
                    "description": "向同名账户转出次数",
                    "type": "integer"
                }
            }
        },
        "rest.ProjectListI": {
            "type": "object",
            "properties": {
                "create_time": {
                    "description": "项目创建时间",
                    "type": "integer"
                },
                "doc_count": {
                    "description": "当前项目下文件总数",
                    "type": "integer"
                },
                "name": {
                    "description": "项目名称",
                    "type": "string"
                },
                "note": {
                    "description": "项目备注",
                    "type": "string"
                },
                "project_id": {
                    "description": "项目id",
                    "type": "integer"
                },
                "type": {
                    "description": "项目类型",
                    "type": "integer"
                }
            }
        },
        "rest.ProjectListResp": {
            "type": "object",
            "properties": {
                "amount_count": {
                    "description": "金额总数 从金额个位开始 三位使用','拼接 例如 123,333,333",
                    "type": "string"
                },
                "count": {
                    "description": "项目总数",
                    "type": "integer"
                },
                "doc_count": {
                    "description": "文档总数",
                    "type": "integer"
                },
                "over_billion": {
                    "description": "金额总数是否超过10亿",
                    "type": "boolean"
                },
                "project_list": {
                    "description": "项目详情",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ProjectListI"
                    }
                },
                "report_count": {
                    "description": "报告总数",
                    "type": "integer"
                },
                "risk_level_1_count": {
                    "description": "高风险项目数",
                    "type": "integer"
                },
                "risk_level_2_count": {
                    "description": "中风险项目数",
                    "type": "integer"
                },
                "risk_level_3_count": {
                    "description": "低风险项目数",
                    "type": "integer"
                }
            }
        },
        "rest.Result1": {
            "type": "object",
            "properties": {
                "data": {},
                "err": {},
                "ret": {
                    "type": "integer"
                }
            }
        },
        "rest.SummaryTypeI": {
            "type": "object",
            "properties": {
                "timer": {
                    "description": "2023-01",
                    "type": "string"
                },
                "type": {
                    "description": "1 摘要完整 2 摘要缺失",
                    "type": "integer"
                }
            }
        },
...
        "rest.ThemeAccountInfo": {
            "type": "object",
            "properties": {
                "associate_info": {
                    "description": "主体账户涉及的银行账户",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.AssociateInfoI"
                    }
                },
                "theme_name": {
                    "description": "主体账户名",
                    "type": "string"
                }
            }
        },
...
    }
}